Adds tons of new things since I haven't pushed anything in a while
Now I have one quadtree takes in the abstract class ObjectBase which EnemyBase, PlayerCharacter and Projectile inherits from.

Enemies are now using steering behaviors like arrive and face which makes them slow down when getting closer to the player and orienting itself so it looks at the player.

The boar is now dashing toward the player when in range instead of doing a similar attack as the human with a sword.

I added two new states in the statestack which are called SurvivalState and TacticalState which inherits from the InGameState. Currently it only run the survival state, which is the same as before. The tactical state is in progress and the idea for it is to have waves that spawn as levels where you are fighting enemies that holds a formation. I have the code for formations done, however I need to spend more time understanding how I can change the formations in a productive way.